requirements.txt
# EDON Gateway Requirements
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
requests>=2.31.0
prometheus-client
python-dotenv
stripe
PyJWT
cryptography

# EDON Guard (local import)
# Assumes edon_demo is in PYTHONPATH or installed
